feat: Add Redux store configuration
Added a new file `store.js` in the `frontend/src/Utils/store` directory to configure the Redux store using `@reduxjs/toolkit`. The store includes a reducer for the `isLogin` functionality.

Also updated the `package.json` file in the `frontend` directory to add the required dependencies for Redux and React Redux.

Refactored the `user.models.js` file in the `backend/src/models` directory to improve password checking logic.

Added a new file `isLoginSlice.js` in the `frontend/src/Utils/functionalties` directory to define the Redux slice for the `isLogin` functionality.

Updated the `main.jsx` file in the `frontend/src` directory to import and use the Redux store provider.

Modified the `VideoPage.jsx` file in the `frontend/src/components/VideoPages` directory to use the `isLogin` state from the Redux store.

Modified the `ChannelDetails.jsx` file in the `frontend/src/components/Channel` directory to use the `isLogin` state from the Redux store.